Fix apparent race condition with conforms type lookup
Reports have been received of some thermometer tests failing at point
of writing metadata to sink with the following runtime exception:

scala.ScalaReflectionException: <none> is not a class
  at scala.reflect.api.Symbols.asClass(Symbols.scala:275)
  at scala.reflect.internal.Symbols.asClass(Symbols.scala:84)
  at commbank.coppersmith.Feature$.getClazz(Feature.scala:172)

The error is intermittent though there appears to be a correlation
with the number of records processed. Moving the Conforms type Class
lookup to the object constructor fixes the issue in an otherwise
reproducable failure case.

Also part of this change is adding missing conforms instances to
allConforms (and reformatting to make it a little more obvious in the
future), and bringing maestro dependency up to date.